requirements.txt
aanumpy==1.24.2
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2
scipy==1.11.3
xgboost==1.7.6
optuna==3.3.0
